I was checking a test app in valgrind and much to my surprise it was complaining about a memleak in libmicrohttpd. In check_argument_match() a buffer is allocated using strdup() but freed nowhere.
I wouldn't have noticed this leak if it wasn't for valgrind because if a URI is requested without any arguments (my usual case)
the buffer that gets allocated has 'only' a length of 1 byte, thus memory usage will build up very slowly over time.

The patch attached fixes it.
Btw. the indentation of that file is a mess, tabs and spaces are mixed :-|

Regards,
Andreas




git-svn-id: https://gnunet.org/svn/libmicrohttpd@35864 140774ce-b5e7-0310-ab8b-a85725594a96